Adding a tutorial for new timing additions, updating for new flake8 v6
