+++ .flake8
@@ -3,15 +3,24 @@ max-line-length = 120
 max-complexity = 45
 ignore =
   E203
-  E225      # missing whitespace around operator
-  E226      # missing whitespace around arithmetic operator
-  W503      # line break before binary operator; conflicts with black
-  W504      # Don't want to change with code frozen...
-  E722      # bare except ok
-  E731      # lambda expressions ok
-  E266      # block comments
-  E741      # ambiguous variable names
-  E501      # lines
+  # missing whitespace around operator
+  E225
+  # missing whitespace around arithmetic operator
+  E226
+  # line break before binary operator; conflicts with black
+  W503
+  # Don't want to change with code frozen...
+  W504
+  # bare except ok
+  E722
+  # lambda expressions ok
+  E731
+  # block comments
+  E266
+  # ambiguous variable names
+  E741
+  # lines
+  E501
 exclude =
     .git
     .tox
